refactor: create a new component for the media capture microphone. revert changes from PVCamera to master.
