+++ Sources/MixedReality/Microsoft.Psi.MixedReality.UniversalWindows/MediaCaptureMicrophoneConfiguration.cs
@@ -1,0 +1,21 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
+namespace Microsoft.Psi.MixedReality
+{
+    /// <summary>
+    /// Configuration for the <see cref="MediaCaptureMicrophone"/> component.
+    /// </summary>
+    public class MediaCaptureMicrophoneConfiguration
+    {
+        /// <summary>
+        /// Gets or sets a value indicating whether the audio is emitted.
+        /// </summary>
+        public bool OutputAudio { get; set; } = true;
+
+        /// <summary>
+        /// Gets or sets the settings for the Microphone.
+        /// </summary>
+        public MicrophoneConfiguration MicrophoneConfiguration { get; set; } = new MicrophoneConfiguration();
+    }
+}
